requirements.txt
praw
nltk
transformers
pandas
sqlalchemy
matplotlib
seaborn
psycopg2-binary
PyYAML